crawlers/xiaohongshu_crawler.py
"""
Xiaohongshu (Little Red Book) crawler implementation
"""
from typing import List, Dict
import requests
from bs4 import BeautifulSoup
import time
from config import Config
from .base_crawler import BaseCrawler


class XiaohongshuCrawler(BaseCrawler):
    """Crawler for Xiaohongshu platform"""
    
    def __init__(self):
        """Initialize Xiaohongshu crawler"""
        super().__init__('Xiaohongshu')
        self.session = requests.Session()
        self.base_url = 'https://www.xiaohongshu.com'
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36',
            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
            'Accept-Language': 'zh-CN,zh;q=0.9,en;q=0.8',
            'Referer': 'https://www.xiaohongshu.com/'
        }
    
    def authenticate(self) -> bool:
        """
        Set up session with cookies
        
        Returns:
            True if setup successful, False otherwise
        """
        try:
            if Config.XIAOHONGSHU_COOKIE:
                # Parse cookie string and add to session
                cookie_dict = {}
                for item in Config.XIAOHONGSHU_COOKIE.split(';'):
                    if '=' in item:
                        key, value = item.strip().split('=', 1)
                        cookie_dict[key] = value
                
                for key, value in cookie_dict.items():
                    self.session.cookies.set(key, value)
                
                self.logger.info("Xiaohongshu session configured with cookies")
            else:
                self.logger.warning("No Xiaohongshu cookie configured, using anonymous access")
            
            return True
            
        except Exception as e:
            self.logger.error(f"Xiaohongshu authentication failed: {e}")
            return False
    
    def get_user_posts(self, username: str, limit: int = 10) -> List[Dict]:
        """
        Get recent posts from a Xiaohongshu user
        
        Note: This is a simplified implementation. Real implementation would need:
        - Proper API endpoint discovery
        - Anti-scraping bypass (captcha, rate limiting)
        - User ID resolution
        
        Args:
            username: Xiaohongshu username or user ID
            limit: Maximum number of posts to fetch
            
        Returns:
            List of post dictionaries
        """
        posts = []
        
        try:
            # Note: This is a placeholder implementation
            # Real implementation would require:
            # 1. Finding the user's profile page URL
            # 2. Extracting user ID
            # 3. Calling the proper API endpoints (which may be protected)
            # 4. Handling anti-scraping measures
            
            user_url = f"{self.base_url}/user/profile/{username}"
            
            # Add delay to avoid rate limiting
            time.sleep(2)
            
            response = self.session.get(
                user_url,
                headers=self.headers,
                timeout=10
            )
            
            if response.status_code != 200:
                self.logger.warning(
                    f"Failed to fetch Xiaohongshu profile for {username}: "
                    f"Status {response.status_code}"
                )
                return posts
            
            # Parse HTML (note: real Xiaohongshu pages are heavily JS-rendered)
            soup = BeautifulSoup(response.text, 'html.parser')
            
            # This is a simplified example - actual implementation would need
            # to handle the real page structure and possibly use Selenium
            # for JavaScript rendering
            
            # For now, return a placeholder indicating the need for proper implementation
            self.logger.warning(
                f"Xiaohongshu crawler needs proper implementation with "
                f"JavaScript rendering (Selenium/Playwright) for user {username}"
            )
            
            # Placeholder post structure
            posts.append({
                'id': 'placeholder',
                'username': username,
                'text': 'Xiaohongshu crawler requires proper implementation with browser automation',
                'created_at': None,
                'note': 'This is a placeholder. Implement with Selenium/Playwright for real data.'
            })
            
        except requests.RequestException as e:
            self.logger.error(f"Network error fetching Xiaohongshu data for {username}: {e}")
        except Exception as e:
            self.logger.error(f"Error fetching Xiaohongshu posts for {username}: {e}")
        
        return posts
    
    def get_user_posts_api(self, user_id: str, limit: int = 10) -> List[Dict]:
        """
        Get posts using Xiaohongshu API (if available)
        
        Args:
            user_id: Xiaohongshu user ID
            limit: Maximum number of posts to fetch
            
        Returns:
            List of post dictionaries
        """
        # This would be the proper API implementation when endpoints are discovered
        # Xiaohongshu's API is not publicly documented and may require reverse engineering
        posts = []
        
        try:
            # Placeholder for API implementation
            api_url = f"{self.base_url}/api/sns/web/v1/user/{user_id}/notes"
            
            params = {
                'num': limit,
                'cursor': ''
            }
            
            time.sleep(2)  # Rate limiting
            
            response = self.session.get(
                api_url,
                headers=self.headers,
                params=params,
                timeout=10
            )
            
            if response.status_code == 200:
                data = response.json()
                # Parse response based on actual API structure
                # This is a placeholder
                self.logger.info(f"API response received for user {user_id}")
            else:
                self.logger.warning(f"API request failed: {response.status_code}")
                
        except Exception as e:
            self.logger.error(f"API error: {e}")
        
        return posts
