requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
schedule==1.2.0
python-dotenv==1.0.0
openai==1.3.0
pymongo==4.6.0
redis==5.0.1
apscheduler==3.10.4
tweepy==4.14.0
lxml==4.9.3
aiohttp==3.9.1
pytest==7.4.3
gunicorn==21.2.0
